Move functions from `bandstructure.contact` to `device.contact`
- PR comment.
