+++ src/quatrex/bandstructure/contact.py
@@ -6,15 +6,10 @@ import numpy as np
 from scipy.optimize import minimize_scalar
 
 from qttools import NDArray, xp
-from qttools.comm import comm
 from qttools.kernels import linalg
-from qttools.toeplitz.circulant import construct_circulant_cell
-from qttools.utils.gpu_utils import get_host
 from quatrex.core.statistics import fermi_dirac
-from quatrex.device import Contact, Device
 from quatrex.electrostatics.geometry_config import Region, VolumeProperties
 from quatrex.electrostatics.meshing import inside_shape
-from quatrex.grid import monkhorst_pack
 
 
 def _real_to_kspace(
@@ -211,191 +206,3 @@ def contact_fermi_level(
     )
 
     return result.x
-
-
-def compute_contact_bandstructure(
-    h_xx: dict,
-    s_xx: dict,
-    kpoint: NDArray,
-    contact: Contact,
-    kpoints_transport: NDArray,
-) -> NDArray:
-    """Slices and expands the inptut matrices, and computes the contact
-    band structure for a given contact at a specific k-point.
-
-    Parameters
-    ----------
-    h_xx : dict
-        The Hamiltonian matrix blocks of a single contact layer. Already
-        summed over k-points.
-    s_xx : dict
-        The overlap matrix blocks of a single contact layer. Already
-        summed over k-points.
-    kpoint : NDArray
-        The k-point at which to compute the band structure.
-    contact : Contact
-        The contact object containing information about the contact.
-    kpoints_transport : NDArray
-        The k-points along the transport direction.
-
-    Returns
-    -------
-    e_k : np.ndarray
-        The eigenvalues for the contact band structure.
-
-    """
-    grid = (contact.transport_repetitions + 1,) + contact.transverse_repetition_grid
-
-    h_xx_tmp = {}
-    s_xx_tmp = {}
-    # shuffle keys to to have natural order a,b,c
-    for i, j, k in np.ndindex(*grid):
-        index = [j, k]
-        index.insert(contact.transport_direction, i)
-        index = tuple(index)
-        h_xx_tmp[index] = h_xx[i, j, k].toarray()
-        s_xx_tmp[index] = s_xx[i, j, k].toarray()
-    h_xx = h_xx_tmp
-    s_xx = s_xx_tmp
-
-    phases = tuple(np.exp(2j * np.pi * k) for k in kpoint)
-    phases = (
-        phases[: contact.transport_direction]
-        + phases[contact.transport_direction + 1 :]
-    )
-
-    H_XX = tuple(
-        construct_circulant_cell(
-            matrix_dict=h_xx,
-            transport_cell_size=contact.transport_repetitions,
-            transport_ind=contact.transport_direction,
-            block_index=block_index,
-            sections=contact.transverse_repetition_grid,
-            phases=phases,
-            key_assumption="half",
-        )
-        for block_index in [-1, 0, 1]
-    )
-    S_XX = tuple(
-        construct_circulant_cell(
-            matrix_dict=s_xx,
-            transport_cell_size=contact.transport_repetitions,
-            transport_ind=contact.transport_direction,
-            block_index=block_index,
-            sections=contact.transverse_repetition_grid,
-            phases=phases,
-            key_assumption="half",
-        )
-        for block_index in [-1, 0, 1]
-    )
-
-    return contact_band_structure(kpoints_transport, H_XX, S_XX)
-
-
-def compute_contact_band_properties(
-    contact: Contact,
-    device: Device,
-) -> tuple[float, float, float]:
-    """Computes the Fermi level for the contact from the Hamiltonian and
-    overlap matrices.
-
-    Parameters
-    ----------
-    contact : Contact
-        The contact object.
-    device : Device
-        The device object.
-
-    Returns
-    -------
-    fermi_level : float
-        The computed Fermi level in eV.
-    mid_gap_energy : float
-        The recomputed mid-gap energy based on the band structure.
-    conduction_band_edge : float
-        The energy of the conduction band edge in eV.
-
-    """
-    contact_config = contact.contact_config
-    device_config = device.device_config
-
-    kpoints_transport = xp.linspace(
-        -xp.pi,
-        xp.pi,
-        contact_config.num_kpoints_transport,
-        endpoint=False,
-    )
-
-    transverse_axes = [0, 1, 2]
-    transverse_axes.remove(contact.transport_direction)
-
-    kpoints = monkhorst_pack(device_config.kpoint_grid, device_config.kpoint_shift)
-
-    e_k = xp.zeros(
-        (
-            len(kpoints_transport),
-            kpoints.shape[0],
-            len(contact.unit_cell_orbital_indices[contact.origin_key])
-            * contact.transport_repetitions
-            * np.prod(contact.transverse_repetition_grid),
-        ),
-        dtype=float,
-    )
-
-    hamiltonians = device.hamiltonians
-    overlaps = device.overlap_matrices
-    for m, kpoint in enumerate(kpoints):
-        h_xx = contact.get_contact_blocks(
-            matrices=hamiltonians,
-            kpoint=kpoint,
-            upper=True,
-        )
-        s_xx = contact.get_contact_blocks(
-            matrices=overlaps,
-            kpoint=kpoint,
-            upper=True,
-        )
-
-        e_k[:, m, :] = compute_contact_bandstructure(
-            h_xx=h_xx,
-            s_xx=s_xx,
-            kpoint=kpoint,
-            contact=contact,
-            kpoints_transport=kpoints_transport,
-        )
-
-    # Average over transverse k-points.
-    e_k = xp.mean(e_k, axis=1)
-
-    doping_density = contact_doping_density(
-        coordinates=get_host(
-            device.orbital_coordinates[
-                contact.unit_cell_orbital_indices[contact.origin_key]
-            ]
-        ),
-        geometry_regions=device_config.geometry.regions,
-    )
-
-    fermi_level = contact_fermi_level(
-        e_k=e_k,
-        kpoints=kpoints_transport,
-        mid_gap_energy=contact.mid_gap_energy,
-        cell_volume=contact.cell_volume,
-        doping_density=doping_density,
-        temperature=contact.temperature,
-    )
-
-    # Recompute the actual mid-gap energy from the band structure.
-    valence_band_edge, conduction_band_edge = contact_band_edges(
-        e_k, contact.mid_gap_energy
-    )
-    mid_gap_energy = 0.5 * (conduction_band_edge + valence_band_edge)
-
-    if comm.rank == 0:
-        print(f"    Doping density: {doping_density} Å^-3", flush=True)
-        print(f"    Fermi level: {fermi_level} eV", flush=True)
-        print(f"    Conduction band minimum: {conduction_band_edge} eV", flush=True)
-        print(f"    Valence band maximum: {valence_band_edge} eV", flush=True)
-        print(f"    Recomputed mid-gap energy: {mid_gap_energy} eV", flush=True)
-
-    return fermi_level, mid_gap_energy, conduction_band_edge
